changed Replay structure to save int values for the angles and changed the ReplayRecorderScript to reproduce replays when the ship was released, not before

+++ superflyingthing-core/src/main/java/com/gemserk/games/superflyingthing/components/Replay.java
@@ -12,9 +12,9 @@ public class Replay {
 		public int time;
 
 		public float x, y;
-		public float angle;
+		public int angle;
 
-		public ReplayEntry(int time, float x, float y, float angle) {
+		public ReplayEntry(int time, float x, float y, int angle) {
 			this.time = time;
 			this.x = x;
 			this.y = y;
